Two new list options: CONFIGURATIONS and OPTIONAL_GROUPS.
CONFIGURATIONS <list> to pass through for which configurations this test is to be valid.

OPTIONAL_GROUPS <list> to specify one or more nams. If any of the specified names (or the special "ALL" name) are to be found in the CMake cache variable CET_TEST_GROUPS, the test will be configured and executed (subject to CONFIGURATIONS, above); otherwise it will not.