Export bin/cmake module directory to Config script
Supply a custom 'ProjectConfig.cmake' script so that the path under
which cetbuildtools programs are installed is set. This allows the
programs to be used by the cetbuildtools CMake scripts without
requiring the PATH to be configured.

Update cet_test to use this method to run cet_exec_test script.